+++ build.rs
@@ -4,13 +4,14 @@ fn main() {
         .atleast_version("6")
         .probe("lhapdf")
         .unwrap();
-    let mut cpp_build = cpp_build::Config::new();
+
+    let mut build = cxx_build::bridge("src/lib.rs");
 
     for include_path in lhapdf.include_paths {
-        cpp_build.include(include_path);
+        build.include(include_path);
     }
 
-    cpp_build.build("src/lib.rs");
+    build.compile("lhapdf-rust-cxx-bridge");
 
     for lib_path in lhapdf.link_paths {
         println!("cargo:rustc-link-search={}", lib_path.to_str().unwrap());
@@ -19,7 +20,18 @@ fn main() {
     for lib in lhapdf.libs {
         println!("cargo:rustc-link-lib={}", lib);
     }
+
+    println!("cargo:rerun-if-changed=src/lib.rs");
+    println!("cargo:rerun-if-changed=include/wrappers.hpp");
 }
 
 #[cfg(feature = "docs-only")]
-fn main() {}
+fn main() {
+    cxx_build::bridge("src/lib.rs")
+        .define("FAKE_WRAPPERS", "1")
+        .compile("lhapdf-rust-cxx-bridge");
+
+    println!("cargo:rerun-if-changed=src/lib.rs");
+    println!("cargo:rerun-if-changed=include/fake-lhapdf.hpp");
+    println!("cargo:rerun-if-changed=include/wrappers.hpp");
+}